Fix wrong file modified
