+++ src/gui/new_item/window.cpp
@@ -1,0 +1,9 @@
+#include "gui/new_item/window.hpp"
+
+namespace Toolbox::UI {
+
+    NewItemWindow::NewItemWindow(const std::string &name) {
+    
+    }
+
+}
